Move the folders and update the modal
