+++ src/views/CoreUnitBudgetStatement/components/ExpenseReport/ExpenseReport.tsx
@@ -1,11 +1,11 @@
 import styled from '@emotion/styled';
-import CategoryModalComponent from '@ses/components/BasicModal/CategoryModalComponent';
 import { CustomLink } from '@ses/components/CustomLink/CustomLink';
 import { ResourceType } from '@ses/core/models/interfaces/types';
 import { MAKER_BURN_LINK } from '@ses/core/utils/const';
 import lightTheme from '@ses/styles/theme/themes';
 import React from 'react';
 import { AdvancedInnerTable } from '@/components/AdvancedInnerTable/AdvancedInnerTable';
+import CategoryModalComponent from '@/components/AdvancedInnerTable/BasicModal/CategoryModalComponent';
 import Container from '@/components/Container/Container';
 import Tabs from '@/components/Tabs/Tabs';
 import {
